refactor: VO 객체 반영으로 인한 수정 - 패키지 경로 변경 - Repository 파라미터 String -> VO

+++ src/main/java/com/smallsquare/modules/user/application/service/UserService.java
@@ -1,10 +1,11 @@
 package com.smallsquare.modules.user.application.service;
 
-import com.smallsquare.common.exception.exception.UserException;
 import com.smallsquare.modules.user.domain.entity.User;
 import com.smallsquare.modules.user.domain.enums.Role;
 import com.smallsquare.modules.user.domain.repository.UserQueryRepository;
 import com.smallsquare.modules.user.domain.repository.UserRepository;
+import com.smallsquare.modules.user.domain.vo.*;
+import com.smallsquare.modules.user.exception.exception.UserException;
 import com.smallsquare.modules.user.infrastructure.jwt.JwtProvider;
 import com.smallsquare.modules.user.infrastructure.jwt.JwtUtil;
 import com.smallsquare.modules.user.infrastructure.redis.RedisService;
@@ -16,9 +17,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.StringUtils;
 
-import static com.smallsquare.common.exception.errorCode.UserErrorCode.*;
+import java.util.Optional;
+
+import static com.smallsquare.modules.user.exception.errorCode.UserErrorCode.*;
 
 @Service
 @RequiredArgsConstructor
@@ -39,18 +41,40 @@ public class UserService {
      */
     @Transactional
     public void signup(UserSignupReqDto reqDto) {
-        // 1. username, nickname, email 중복검사
-        validateDuplicate(reqDto);
 
-        // 2. 비밀번호와 비밀번호 확인이 일치하는지 확인
-        validatePasswordMatch(reqDto.getPassword(), reqDto.getCheckPassword());
+        // 1. 비밀번호와 비밀번호 확인이 일치하는지 확인 (평문)
+        validateSamePassword(reqDto.getPassword(), reqDto.getCheckPassword());
 
-        // 3. 이메일 인증이 이루어졌는지 확인
+        // 2. 이메일 인증이 이루어졌는지 확인
         // TODO: PostMan 테스트 시 너무 불편해서 주석처리 한 다음 실제 서비스할 때는 주석 제거
-        //verifyEmail(reqDto.getEmail());
+        // verifyEmail(reqDto.getEmail());
 
-        // 3. DTO의 필드와 비밀번호를 인코딩해서 User 객체로 변환 후 저장
-        User user = User.of(reqDto, passwordEncoder.encode(reqDto.getPassword()));
+        // 3. 비밀번호 객체를 만들어 인코딩 후 대입
+        Password password = new Password(reqDto.getPassword(), passwordEncoder);
+
+        // 4. 이메일 객체를 만들어 중복검사
+        Email email = new Email(reqDto.getEmail());
+        if (userRepository.existsByEmail(email)) {
+            throw new UserException(DUPLICATED_EMAIL);
+        }
+
+        // 5. username 객체 만들어 중복 검사
+        Username username = new Username(reqDto.getUsername());
+        if (userRepository.existsByUsername(username)) {
+            throw new UserException(DUPLICATED_USERNAME);
+        }
+
+        // 6. nickname 객체를 만들어 중복검사
+        Nickname nickname = new Nickname(reqDto.getNickname());
+        if (userRepository.existsByNickname(nickname)) {
+            throw new UserException(DUPLICATED_NICKNAME);
+        }
+
+        // 7. name 객체 생성
+        Name name = new Name(reqDto.getName());
+
+        // 8. DTO의 필드 User 객체로 변환 후 저장
+        User user = User.of(username, password, email, nickname, name);
         userRepository.save(user);
     }
 
@@ -62,20 +86,25 @@ public class UserService {
     @Transactional
     public UserLoginResDto login(UserLoginReqDto reqDto) {
 
-        // 1. 유저조회
-        User user = userRepository.findByUsername(reqDto.getUsername())
-                .orElseThrow(() -> new UserException(USER_NOT_FOUND));
+        // 1. username 객체 생성
+        Username username = new Username(reqDto.getUsername());
 
-        // 2. 비밀번호가 입력한 것과 DB의 비밀번호가 일치하는지 확인
-        validatePassword(reqDto.getPassword(), user.getPassword());
+        // 2. 유저조회
+        User user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new UserException(USER_NOT_FOUND));
 
         // 3. 탈퇴한 유저인지 검증
         if (!user.getIsActive()) {
             throw new UserException(INACTIVE_ACCOUNT);
         }
 
-        // 4. 토큰 생성
-        JwtTokenReqDto tokenReqDto = createTokenDto(user);
+        // 4. 비밀번호가 입력한 것과 DB의 비밀번호가 일치하는지 확인
+        if (!user.getPassword().matchPassword(reqDto.getPassword(), passwordEncoder)) {
+            throw new UserException(PASSWORD_NOT_MATCHED);
+        }
+
+        // 5. 토큰 생성
+        JwtTokenReqDto tokenReqDto = jwtUtil.createTokenDto(user);
 
         String accessToken = jwtProvider.createAccessToken(tokenReqDto);
         String refreshToken = jwtProvider.createRefreshToken(tokenReqDto);
@@ -124,15 +153,40 @@ public class UserService {
      */
     @Transactional
     public UserUpdateResDto updateUserInfo(Long userId, UserUpdateReqDto reqDto) {
-        User user = userRepository.findById(userId).orElseThrow(() -> new UserException(USER_NOT_FOUND)); // db에서 데이터를 가져와서 1차 캐시에 저장 (영속상태)
-        // 영속 객체의 필드 값 변경 (더티 체킹 발생) -> Transactional 범위가 끝날 때 변경 감지 -> 감지되면 update 쿼리 실행
-        user.updateInfo(reqDto);
 
+        // 1. 해당 회원 조회
+        User user = userRepository.findById(userId).orElseThrow(() -> new UserException(USER_NOT_FOUND)); // db에서 데이터를 가져와서 1차 캐시에 저장 (영속상태)
+
+        // 2. 이메일 객체를 만들어 중복검사
+        Email email = new Email(reqDto.getEmail());
+        if (userRepository.existsByEmail(email) && !user.getEmail().equals(email)) {
+            throw new UserException(DUPLICATED_EMAIL);
+        }
+
+        // 3. username 객체 만들어 중복 검사
+        Username username = new Username(reqDto.getUsername());
+        if (userRepository.existsByUsername(username) && !user.getUsername().equals(username)) {
+            throw new UserException(DUPLICATED_USERNAME);
+        }
+
+        // 4. nickname 객체를 만들어 중복검사
+        Nickname nickname = new Nickname(reqDto.getNickname());
+        if (userRepository.existsByNickname(nickname) && !user.getNickname().equals(nickname)) {
+            throw new UserException(DUPLICATED_NICKNAME);
+        }
+
+        // 5. Name 객체 생성
+        Name name = new Name(reqDto.getName());
+
+        // 6. 영속 객체의 필드 값 변경 (더티 체킹 발생) -> Transactional 범위가 끝날 때 변경 감지 -> 감지되면 update 쿼리 실행
+        user.updateInfo(reqDto, username, email, nickname, name);
+
+        // 7. 반환
         UserUpdateResDto resDto = UserUpdateResDto.builder()
-                .username(user.getUsername())
-                .name(user.getName())
-                .email(user.getEmail())
-                .nickname(user.getNickname())
+                .username(user.getUsername().getUsername())
+                .name(user.getName().getName())
+                .email(user.getEmail().getEmail())
+                .nickname(user.getNickname().getNickname())
                 .build();
 
         return resDto;
@@ -148,7 +202,9 @@ public class UserService {
         User user = userRepository.findById(userId).orElseThrow(() -> new UserException(USER_NOT_FOUND));
 
         // 1. 비밀번호가 일치하는지 확인
-        validatePassword(reqDto.getPassword(), user.getPassword());
+        if (!user.getPassword().matchPassword(reqDto.getPassword(), passwordEncoder)) {
+            throw new UserException(PASSWORD_NOT_MATCHED);
+        }
 
         // 2. User 테이블의 isActive를 false로 변경
         user.deactivate();
@@ -159,21 +215,25 @@ public class UserService {
     public void findAndChangePassword(UserFindPasswordReqDto reqDto) {
 
         // 1. 비밀번호 일치 여부 확인
-        validatePasswordMatch(reqDto.getPassword(), reqDto.getCheckPassword());
+        validateSamePassword(reqDto.getPassword(), reqDto.getCheckPassword());
 
         // 2. redis키로 이메일을 추출
         String redisKey = "findPassword:token:" + reqDto.getPasswordToken();
-        String email = redisService.get(redisKey);
+        String emailKey = redisService.get(redisKey);
+
+        // 3. Email 객체 생성
+        Email email = new Email(emailKey);
 
         // 3. 해당 email을 가지고 있는 User 조회
         User user = userRepository.findByEmail(email).orElseThrow(() -> new UserException(USER_NOT_FOUND));
 
         // 4. 이전 비밀번호와 동일하다면 예외 발생
-        validateNewPasswordIsNotSameAsOld(reqDto.getPassword(), user.getPassword());
+        if (user.getPassword().isSamePassword(reqDto.getPassword(), passwordEncoder)) {
+            throw new UserException(SAME_AS_OLD_PASSWORD);
+        }
 
-        // 5. 비밀번호 업데이트
-        String newPassword = passwordEncoder.encode(reqDto.getPassword());
-        user.updatePassword(newPassword);
+        // 5. 비밀번호 변경
+        user.updatePassword(reqDto.getPassword(), passwordEncoder);
 
         // 6. redis에 해당 키 삭제
         redisService.delete(redisKey);
@@ -185,25 +245,25 @@ public class UserService {
      * @param userId
      * @param reqDto
      */
-
     @Transactional
     public void changePassword(Long userId, UserPasswordChangeReqDto reqDto) {
 
         // 1. 입력한 비밀번호가 DB의 비밀번호와 일치하는지 확인
         User user = userRepository.findById(userId).orElseThrow(() -> new UserException(USER_NOT_FOUND));
-        validatePassword(reqDto.getPassword(), user.getPassword());
+        if (!user.getPassword().matchPassword(reqDto.getPassword(), passwordEncoder)) {
+            throw new UserException(PASSWORD_NOT_MATCHED);
+        }
 
         // 2. 새로 입력한 비밀번호와 비밀번호 확인이 서로 일치하는지 확인
-        validatePasswordMatch(reqDto.getNewPassword(), reqDto.getCheckNewPassword());
+        validateSamePassword(reqDto.getNewPassword(), reqDto.getCheckNewPassword());
 
         // 3. 기존 비밀번호와 새로 입력한 비밀번호가 같은 경우
-        validateNewPasswordIsNotSameAsOld(reqDto.getNewPassword(), user.getPassword());
+        if (user.getPassword().isSamePassword(reqDto.getNewPassword(), passwordEncoder)) {
+            throw new UserException(SAME_AS_OLD_PASSWORD);
+        }
 
-        // 4. 비밀번호 인코딩
-        String newPassword = passwordEncoder.encode(reqDto.getNewPassword());
-
-        // 5. 유저의 비밀번호를 새로운 비밀번호로 변경
-        user.updatePassword(newPassword);
+        // 4. 유저의 비밀번호를 새로운 비밀번호로 변경
+        user.updatePassword(reqDto.getNewPassword(), passwordEncoder);
     }
 
     /**
@@ -218,25 +278,24 @@ public class UserService {
         jwtUtil.validateToken(refreshToken);
 
         // 2. RefreshToken이 블랙리스트인지 확인
-        validateRefreshTokenBlackList(refreshToken);
+        jwtUtil.validateRefreshTokenBlackList(refreshToken);
 
-        // 3. 사용자 정보 추출
-        User user = User.builder()
-                .id(jwtUtil.getUserId(refreshToken))
-                .username(jwtUtil.getUsername(refreshToken))
-                .name(jwtUtil.getName(refreshToken))
-                .email(jwtUtil.getEmail(refreshToken))
-                .nickname(jwtUtil.getNickname(refreshToken))
-                .role(Role.valueOf(jwtUtil.getRole(refreshToken)))
-                .build();
+        // 3. user 조회
+        User user = userRepository.findById(jwtUtil.getUserId(refreshToken))
+                .orElseThrow(() -> new UserException(USER_NOT_FOUND));
 
-        // 4. 사용된 리프레시 토큰을 블랙리스트로 저장
+        // 4. 탈퇴한 사용자인지 검증
+        if (user.getIsActive() == false) {
+            throw new UserException(INACTIVE_ACCOUNT);
+        }
+
+        // 5. 사용된 리프레시 토큰을 블랙리스트로 저장
         redisService.saveRefreshBlackList(refreshToken, jwtUtil.getRemainingExpirationMillis(refreshToken));
 
-        // 5. Token에 사용자 정보를 담기
-        JwtTokenReqDto userInfo = createTokenDto(user);
+        // 6. Token에 사용자 정보를 담기
+        JwtTokenReqDto userInfo = jwtUtil.createTokenDto(user);
 
-        // 6. 새로운 엑세스 토큰과 리프레시 토큰을 만들고 반환
+        // 7. 새로운 엑세스 토큰과 리프레시 토큰을 만들고 반환
         String newAccessToken = jwtProvider.createAccessToken(userInfo);
         String newRefreshToken = jwtProvider.createRefreshToken(userInfo);
 
@@ -252,92 +311,25 @@ public class UserService {
     // ==================== 검증 & 편의 메소드 ==================== //
 
     /**
-     * username, nickname, email 중복검사
-     * @param reqDto
+     * 비밀번호와 비밀번호 확인이 일치하는지 검증
+     * @param pw1
+     * @param pw2
      */
-    private void validateDuplicate(UserSignupReqDto reqDto) {
-        if (userRepository.existsByUsername(reqDto.getUsername())) {
-            throw new UserException(DUPLICATED_USERNAME);
-        }
-        if (userRepository.existsByEmail(reqDto.getEmail())) {
-            throw new UserException(DUPLICATED_EMAIL);
-        }
-        if (userRepository.existsByNickname(reqDto.getNickname())) {
-            throw new UserException(DUPLICATED_NICKNAME);
-        }
-    }
-
-    /**
-     * 입력한 비밀번호와 DB의 비밀번호가 서로 일치하는지 확인
-     * @note mathes()는 인코딩 안된 평문 비밀번호와 인코딩된 비밀번호를 비교하도록 설계된 메소드
-     * @param password
-     * @param inputPassword
-     */
-    private void validatePassword(String password, String inputPassword) {
-        if (!passwordEncoder.matches(password, inputPassword)) {
-            throw new UserException(PASSWORD_NOT_MATCHED);
-        }
-    }
-
-    /**
-     * TokenDto로 변환하는 메소드
-     * @param user
-     * @return
-     */
-    private JwtTokenReqDto createTokenDto(User user) {
-        return JwtTokenReqDto.builder()
-                .userId(user.getId())
-                .username(user.getUsername())
-                .nickname(user.getNickname())
-                .email(user.getEmail())
-                .name(user.getName())
-                .role(user.getRole())
-                .build();
-    }
-
-    /**
-     * 비밀번호 & 비밀번호 확인이 서로 일치하는지 확인하는 로직
-     * @param password, checkPassword
-     */
-    private void validatePasswordMatch(String password, String checkPassword) {
-        if (!password.equals(checkPassword)) {
+    private void validateSamePassword(String pw1, String pw2) {
+        if (!pw1.equals(pw2)) {
             throw new UserException(PASSWORD_MISMATCH);
         }
     }
 
     /**
-     * 새로 변경된 비밀번호가 이전 비밀번호과 동일한지 확인하는 로직
-     * @param newPassword
-     * @param oldPassword
-     */
-
-    private void validateNewPasswordIsNotSameAsOld(String newPassword, String oldPassword) {
-        if (passwordEncoder.matches(newPassword, oldPassword)) {
-            throw new UserException(SAME_AS_OLD_PASSWORD);
-        }
-    }
-
-    /**
-     * 이메일 인증응 확인하는 로직
+     * 이메일 인증을 확인하는 로직
      * @param email
      * 개발 단계에서는 잠시 미사용 (회원 가입 시 매번 redis에 값을 넣어줘야해서 번거로움)
      */
-
     private void verifyEmail(String email) {
         String status = redisService.get("verifyEmail:email:" + email);
-        if(!status.equals("true")) {
+        if(!"true".equals(status)) {
             throw new UserException(EMAIL_NOT_VERIFIED);
-        }
-    }
-
-    /**
-     * 토큰 재발급 시 기존 폐기된 리프레시 토큰이 사용되었는지 확인하는 로직
-     * @param refreshToken
-     */
-    private void validateRefreshTokenBlackList(String refreshToken) {
-        String key = redisService.get("blacklist:refresh:" + refreshToken);
-        if (StringUtils.hasText(key)) {
-            throw new UserException(EXPIRED_REFRESH_TOKEN);
         }
     }
